backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pydantic
python-dotenv
supabase
joblib